feat: PhotoController #23

+++ MarsBackEnd/Domain/Entity/Profile.cs
@@ -10,7 +10,7 @@ namespace Domain.Entity
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int Id { get; set; }
-        public byte Age { get; set; }
+        public int Age { get; set; }
         public string Address { get; set; }
         
     }
